[Alert activations] Fix partitioning for day and month (#6936)
Jira - https://iguazio.atlassian.net/browse/ML-8768

the problem was that DAYOFMONTH and MONTH (probably, wasn't fully
tested) partition expressions weren't entirely correct. As when we
create partitions we create them with condition ` VALUES LESS THAN
({next_partition_value}`.

For `YEARWEEK` expression, the real generated value is:
```
mysql> select yearweek(now());
+-----------------+
| yearweek(now()) |
+-----------------+
|          202450 |
+-----------------+
```

When for dayofmonth():
```
mysql> select dayofmonth(now());
+-------------------+
| dayofmonth(now()) |
+-------------------+
|                17 |
+-------------------+
```

it is a single number, which means that the records will be always put
to the oldest partition (because any day of the month 1-31 is always
less than next_parititon_value).

This PR changes the expressions from dayofmonth or month to this for
`DAY`:
```
mysql> select YEAR(NOW())*10000 + MONTH(NOW())*100 + DAY(NOW());
+---------------------------------------------------+
| YEAR(NOW())*10000 + MONTH(NOW())*100 + DAY(NOW()) |
+---------------------------------------------------+
|                                          20241217 |
+---------------------------------------------------+
```
And this for `month`:

```
SELECT YEAR('2025-01-01') * 100 + MONTH('2025-01-01');
+------------------------------------------------+
| YEAR('2025-01-01') * 100 + MONTH('2025-01-01') |
+------------------------------------------------+
|                                         202501 |
+------------------------------------------------+
```

(*) Also fixed PartitionInterval to be more general and do not hardcode
`activation_time` as a column name.