[DB] Handling "Out of Range" Error When Committing Values Exceeding Column Limits (#7046)
When committing to the database, if a field value exceeds the allowed
range (for example, trying to commit a value of 144466666666666 to the
current_page field, which is defined as an integer), the database fails
with the following error:
```
DataError('(pymysql.err.DataError) (1264, "Out of range value for column \'current_page\' at row 1")').
```
In this case, an error is raised indicating that the column
'current_page' exceeds the allowed range.

https://iguazio.atlassian.net/browse/ML-8442