[Alerts] Improvements to increase stability and performance (#7048)
Within this pr:
1.There is a risk of encountering issues when storing alert
configurations during an alert service restart. To address this, ensure
that the state is always present when storing an alert configuration.
2. Added a with_state flag to enable fetching both the configuration and
state in a single query.
3. Fixed the unique constraint for the alert_states table. Specifically,
ensuring that parent_id is unique by adding a unique constraint on it
and removing the unnecessary constraint.

jira -  https://iguazio.atlassian.net/browse/ML-8925

---------

Co-authored-by: TomerShor <90552140+TomerShor@users.noreply.github.com>