[Model Monitoring] Fix default `sampling_percentage` on serving deployment when monitoring is disabled  (#7053)
When a user deploys a serving func without mm, the model endpoint is
generated with `monitoring mode` set to `disabled`. However, in this
case, the serving func has no `sampling_percentage` configuration, which
led to a bug during mep creation. This PR fixes this issue by setting a
default `sampling_percentage` value of 100 when it is not specified in
the serving spec.

https://iguazio.atlassian.net/browse/ML-8997