[Alerts] Exclude updated field from alerts response for old client (#7090)
The updated field was added to 1.8.0 but clients using version 1.7.1 do
not expect this field.
As a result, calling `list_alerts_configs` or `get_alert_config` with a
1.7.1 client causes a TypeError due to the unexpected `updated` field
being included in the response.

This PR introduces the logic to exclude the `updated` field from the
alert responses when client version is < 1.8.0,
This was done by using FastAPI's `response_model_exclude_none=True` to
ensure that `None` values are excluded from the response.

Jira:
https://iguazio.atlassian.net/browse/ML-8837

Also a minor refactor was done in this PR: changed the name of
get_enriched_alert to get_alert ( as the older get_alert method was not
used )