[Notifications] Support `end_time` w/ Backwards Compatible Clients (#7273)
https://iguazio.atlassian.net/browse/ML-9294

My previous [PR](https://github.com/mlrun/mlrun/pull/7264) dealt with
the server side run monitor setting end time on runs it sees as phasing
into a terminal state.
However, when running a job that completes very quickly, it is the
client that triggers the run state update before the run monitor gets to
it. Old clients (pre 1.8.0) aren't aware of the `end_time` field and
therefore do not send that field as part of the necessary updates.

Modified `update_run` to check heuristically whether it needs to set the
end time, either from the client's request if given or by itself if the
run is now terminal and does not already have an `end_time`.